+++ src/Products/PloneMeeting/MeetingItem.py
@@ -92,6 +92,7 @@ from Products.PloneMeeting.utils import _storedItemNumber_to_itemNumber
 from Products.PloneMeeting.utils import add_wf_history_action
 from Products.PloneMeeting.utils import addDataChange
 from Products.PloneMeeting.utils import AdvicesUpdatedEvent
+from Products.PloneMeeting.utils import checkMayQuickEdit
 from Products.PloneMeeting.utils import cleanMemoize
 from Products.PloneMeeting.utils import compute_item_roles_to_assign_to_suffixes
 from Products.PloneMeeting.utils import decodeDelayAwareId
@@ -115,7 +116,7 @@ from Products.PloneMeeting.utils import notifyModifiedAndReindex
 from Products.PloneMeeting.utils import rememberPreviousData
 from Products.PloneMeeting.utils import sendMail
 from Products.PloneMeeting.utils import sendMailIfRelevant
-from Products.PloneMeeting.utils import setFieldFromAjax
+from Products.PloneMeeting.utils import set_field_from_ajax
 from Products.PloneMeeting.utils import toHTMLStrikedContent
 from Products.PloneMeeting.utils import transformAllRichTextFields
 from Products.PloneMeeting.utils import updateAnnexesAccess
@@ -3955,23 +3956,6 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
             unrestrictedRemoveGivenObject(item)
             return True
 
-    def _checkMayQuickEdit(self,
-                           bypassWritePermissionCheck=False,
-                           permission=ModifyPortalContent,
-                           expression='',
-                           onlyForManagers=False):
-        """ """
-        tool = api.portal.get_tool('portal_plonemeeting')
-        member = api.user.get_current()
-        res = False
-        if (not onlyForManagers or (onlyForManagers and tool.isManager(self))) and \
-           (bypassWritePermissionCheck or member.has_permission(permission, self)) and \
-           (_evaluateExpression(self, expression)) and \
-           (not (self.hasMeeting() and self.getMeeting().queryState() in Meeting.meetingClosedStates) or
-                tool.isManager(self, realManagers=True)):
-            res = True
-        return res
-
     security.declarePublic('mayQuickEdit')
 
     def mayQuickEdit(self, fieldName, bypassWritePermissionCheck=False, onlyForManagers=False):
@@ -3982,7 +3966,8 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
            a real Manager (Site Administrator/Manager).
            If p_bypassWritePermissionCheck is True, we will not check for write_permission.'''
         field = self.Schema()[fieldName]
-        return self._checkMayQuickEdit(
+        return checkMayQuickEdit(
+            self,
             bypassWritePermissionCheck=bypassWritePermissionCheck,
             permission=field.write_permission,
             expression=self.Schema()[fieldName].widget.condition,
@@ -3996,12 +3981,6 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
     def mayQuickEditItemSignatures(self):
         """Show edit icon if itemSignatures field editable."""
         return self.mayQuickEdit('itemSignatures', bypassWritePermissionCheck=True, onlyForManagers=True)
-
-    security.declareProtected(ModifyPortalContent, 'transformRichTextField')
-
-    def transformRichTextField(self, fieldName, richContent):
-        '''See doc in interfaces.py.'''
-        return richContent
 
     security.declareProtected(ModifyPortalContent, 'onEdit')
 
@@ -6792,7 +6771,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
             self.updateLocalRoles(invalidate=True)
         # versionate given advices if necessary
         self._versionateAdvicesOnItemEdit()
-        return setFieldFromAjax(self, fieldName, fieldValue)
+        return set_field_from_ajax(self, fieldName, fieldValue)
 
     security.declarePublic('getFieldVersion')
 
@@ -6823,8 +6802,8 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
     def _mayChangeAttendees(self):
         """Check that user may quickEdit
            itemAbsents/itemExcused/itemNonAttendees/votes/..."""
-        return self.hasMeeting() and self._checkMayQuickEdit(
-            bypassWritePermissionCheck=True, onlyForManagers=True)
+        return self.hasMeeting() and checkMayQuickEdit(
+            self, bypassWritePermissionCheck=True, onlyForManagers=True)
 
     def _voteIsDeletable(self, vote_number):
         """ """